requirements.txt
flask>=2.0.0
flake8>=6.0.0
black>=23.0.0
bandit>=1.7.0
jsonschema>=4.0.0
pytest>=7.0.0
pytest-flask>=1.2.0
gunicorn>=20.1.0
python-dotenv>=0.19.0 